LayoutTests/http/tests/security/contentSecurityPolicy/report-only-from-header-expected.txt

CONSOLE MESSAGE: [Report Only] Refused to execute inline script because of Content-Security-Policy.

ALERT: PASS
CSP report received:
CONTENT_TYPE: application/x-www-form-urlencoded
HTTP_REFERER: http://127.0.0.1:8000/security/contentSecurityPolicy/report-only-from-header.php
REQUEST_METHOD: POST
=== POST DATA ===
document-url: http://127.0.0.1:8000/security/contentSecurityPolicy/report-only-from-header.php
violated-directive: script-src 'self'
